pyproject.toml
[project]
name = "osmsg"
version = "0.4.0"
description = "OpenStreetMap Stats Generator: Commandline"
readme = "README.md"
authors = [
    {name = "Kshitij Raj Sharma", email = "skshitizraj@gmail.com"},
]
license = {text = "GPL-3.0-only"}
requires-python = ">=3.12"
dependencies = [
    "duckdb",
    "osmium",
    "platformdirs",
    "pyarrow",
    "pydantic",
    "python-dotenv",
    "pytz",         # transitive runtime requirement of duckdb (TIMESTAMPTZ → datetime)
    "requests",
    "shapely",
    "typer",
    "typer-config[yaml]",
]

keywords = ["osm", "stats", "commandline", "openstreetmap"]
classifiers = [
    "Topic :: Utilities",
    "Topic :: Scientific/Engineering :: GIS",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
]

[project.urls]
homepage = "https://github.com/osgeonepal/osmsg"
documentation = "https://github.com/osgeonepal/osmsg"
repository = "https://github.com/osgeonepal/osmsg"

[project.optional-dependencies]
docs = [
    "mkdocs>=1.5.3",
    "mkdocs-material>=9.4.2",
    "mkdocstrings-python>=1.7.0",
    "mkdocs-exclude>=1.0.2",
]

[project.scripts]
osmsg = "osmsg.cli:app"

[build-system]
requires = ["uv_build>=0.5.15,<0.9"]
build-backend = "uv_build"

[tool.uv.build-backend]
module-name = "osmsg"
module-root = ""

[dependency-groups]
dev = [
    "commitizen",
    "pre-commit",
    "ruff",
    "ty",
    "pytest>=7.4.2",
    "pytest-mock>=3.14.0",
    "coverage>=7.6.1",
]

[tool.pytest.ini_options]
addopts = "-ra -q --strict-markers"
testpaths = ["tests"]
markers = [
    "network: tests that hit the public internet (deselect with -m 'not network').",
]

[tool.commitizen]
name = "cz_conventional_commits"
tag_format = "v$version"
version_scheme = "semver"
version_provider = "pep621"
update_changelog_on_bump = true
version_files = [
    "pyproject.toml:version",
    "osmsg/__version__.py",
]

[tool.ruff]
fix = true
line-length = 120
target-version = "py312"
exclude = [
    ".git",
    ".ruff_cache",
    ".vscode",
    "__pypackages__",
    "build",
    "dist",
    "osmsg/__version__.py",
]

[tool.ruff.lint]
select = ["I", "E", "W", "B", "F", "N", "Q", "UP", "SIM"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.ty.src]
include = ["osmsg"]

[tool.coverage.run]
source = ["osmsg"]

[tool.coverage.report]
show_missing = true
